+++ tests/testthat/test-dictionary.R
@@ -57,11 +57,6 @@ test_that("vec_duplicate_any returns single TRUE/FALSE", {
   expect_true(vec_duplicate_any(c(1:10, 1)))
 })
 
-test_that("vec_duplicate_id gives position of first found", {
-  x <- c(1, 2, 3, 1, 4)
-  expect_equal(vec_duplicate_id(x), c(1, 2, 3, 1, 5))
-})
-
 test_that("vec_duplicate_loc is the complement of vec_unique_loc", {
   x <- c(1, 1, 2, 3, 4, 4)
   expect <- vec_seq_along(x)[-vec_unique_loc(x)]
@@ -147,7 +142,6 @@ test_that("duplicate functions take the equality proxy recursively", {
 
   expect_equal(vec_duplicate_any(df), TRUE)
   expect_equal(vec_duplicate_detect(df), c(TRUE, TRUE, FALSE))
-  expect_equal(vec_duplicate_id(df), c(1, 1, 3))
   expect_equal(vec_duplicate_loc(df), 2)
 })
 
@@ -176,7 +170,6 @@ test_that("unique functions can handle scalar types in lists", {
 test_that("duplicate functions works with different encodings", {
   encs <- encodings()
 
-  expect_equal(vec_duplicate_id(encs), rep(1, 3))
   expect_equal(vec_duplicate_detect(encs), rep(TRUE, 3))
   expect_equal(vec_duplicate_any(encs), TRUE)
   expect_equal(vec_duplicate_loc(encs), c(2, 3))
@@ -200,6 +193,26 @@ test_that("vec_unique() works with glm objects (#643)", {
   # class(model$family$initialize) == "expression"
   model <- glm(mpg ~ wt, data = mtcars)
   expect_equal(vec_unique(list(model, model)), list(model))
+})
+
+# first location ----------------------------------------------------------
+
+test_that("vec_first_loc gives the location of first occurrence found", {
+  x <- c(1, 2, 3, 1, 4)
+  expect_equal(vec_first_loc(x), c(1, 2, 3, 1, 5))
+})
+
+test_that("vec_first_loc takes the equality proxy recursively", {
+  local_comparable_tuple()
+
+  x <- tuple(c(1, 1, 2), 1:3)
+  df <- data_frame(x = x)
+
+  expect_equal(vec_first_loc(df), c(1, 1, 3))
+})
+
+test_that("vec_first_loc works with different encodings", {
+  expect_equal(vec_first_loc(encodings()), rep(1, 3))
 })
 
 # matching ----------------------------------------------------------------
@@ -234,7 +247,6 @@ test_that("matching functions take the equality proxy (#375)", {
   expect_identical(unique(x), tuple(c(1, 2), 1:2))
 
   expect_true(vec_duplicate_any(x))
-  expect_identical(vec_duplicate_id(x), c(1L, 2L, 1L))
   expect_identical(vec_unique_count(x), 2L)
 
   expect_identical(vec_duplicate_detect(x), c(TRUE, FALSE, TRUE))